+++ packages/munsell/src/rhodoneaFolios.js
@@ -4,30 +4,28 @@ import * as pol          from '@aryth/polar'
 import { PetalNote }     from '@aryth/polar'
 import * as conv         from '@palett/convert'
 import { iterate }       from '@vect/vector-mapper'
-import { Cuvette }       from './Cuvette.js'
-import { Domain }        from './Domain.js'
+import { Cova }          from './Cova.js'
 
 const { PI, pow, abs, round } = Math
 
-// list<(string hex, string name)>
+// entries<(string hex, string name)>
 export function rhodoneaFolios(
   hsl,
   {
     petals,
     density = 0.01,
     lightMinimum = 0,
-    saturTolerance = 18,
-    domain = Domain.fashion
+    saturTolerance = 18
   }
 ) {
-  const cuvette = Cuvette.select(domain)
   const polarMark = hsl.toPolar()
-  const hexToHsl = cuvette.hexToHsl.slice() // create shallow copy
+  const hexToHsl = Cova.hexToHsl.slice() // create shallow copy
   const saturInterval = Bound.build(hsl.s - saturTolerance, hsl.s + saturTolerance)
   const minL = lightMinimum
   const area = PI * pow(polarMark.r, 2) * (petals % 2 === 0 ? 0.5 : 0.25)
   const maximum = round(density * area)
-  const thresholdPerPhase = maximum / petals
+  const phaseThres = maximum / petals
+  // console.log({ hsl, polarMark, petals })
   const petalNote = PetalNote.build(polarMark.θ, petals)
   const petalCache = {}
   for (let i = 1; i <= petals; i++) { petalCache[i] = [] }
@@ -35,16 +33,16 @@ export function rhodoneaFolios(
   const sortList = []
   const hslIndicator = (hsl) => hsl.h * 10000 + hsl.s * 100 + hsl.l
   let i = 0
-  for (const [hex, hsl] of finiteFlopper(hexToHsl)) {
+  for (const [ hex, hsl ] of finiteFlopper(hexToHsl)) {
     i++
-    let [t, s, r] = hsl
+    let [ t, s, r ] = hsl
     if (r < minL) continue
     if (polarMark.foliateRadius(t, petals) < r) continue
     const phase = petalNote.phase(t)
-    if (thresholdPerPhase <= petalNote.counter[phase]) continue
+    if (phaseThres <= petalNote.counter[phase]) continue
     if (saturInterval.has(s)) {
       petalNote.notePhase(phase)
-      sortList.push([hslIndicator(hsl), hex])
+      sortList.push([ hslIndicator(hsl), hex ])
     } else {
       const ds = abs(hsl.s - s)
       const dr = abs(hsl.l - r)
@@ -53,16 +51,16 @@ export function rhodoneaFolios(
     }
     if (maximum <= petalNote.sum) break
   }
-  iterate(petalNote.counter, ([phase, count]) => {
-    if (count < thresholdPerPhase) {
+  iterate(petalNote.counter, ([ phase, count ]) => {
+    if (count < phaseThres) {
       const cache = petalCache[phase]
       if (cache?.length) {
-        for (let hex of cache.values.slice(0, thresholdPerPhase - count)) {
-          sortList.push([hslIndicator(conv.hexToHsl(hex)), hex])
+        for (let hex of cache.values.slice(0, phaseThres - count)) {
+          sortList.push([ hslIndicator(conv.hexToHsl(hex)), hex ])
         }
       }
     }
   })
-  sortList.sort(([a,], [b,]) => a - b)
-  return sortList.map(([_, hex]) => [hex, cuvette.name(hex)])
+  sortList.sort(([ a ], [ b ]) => a - b)
+  return sortList.map(([ _, hex ]) => [ hex, Cova.name(hex) ])
 }