+++ packages/munsell/test/hsl.properties.unit.js
@@ -1,10 +1,13 @@
-import { HSL }          from '../src/HSL.js'
-import { fluo }         from '@palett/fluo'
 import { hslToRgb }     from '@palett/convert'
+import { fluo }         from '@palett/fluo'
 import { deco, logger } from '@spare/logger'
+import { test } from 'node:test'
+import { HSL }  from '../src/extends/HSL.js'
 
 const hsl = HSL.fromRgb([ 92, 92, 255 ])
 
-logger(`[h] (${hsl.h}) [s] (${hsl.s}) [l] (${hsl.l})`)
-logger(deco(hsl), deco(hsl.toRgb()))
-logger(fluo(hsl.toString(), hslToRgb(hsl)))
+test('hsl properties', () => {
+  logger(`[h] (${hsl.h}) [s] (${hsl.s}) [l] (${hsl.l})`)
+  logger(deco(hsl), deco(hsl.toRgb()))
+  logger(fluo(hsl.toString(), hslToRgb(hsl)))
+})