+++ packages/munsell/test/terminal.crostab.unit.js
@@ -6,10 +6,10 @@ import { DecoCrostab, logger } from '@spare/logger'
 import { Preset }              from '@palett/presets'
 import { Fluo }                from '@palett/fluo'
 import { toneHex }             from '@palett/color-algebra'
-import { indexedTo }           from '@vect/object-mapper'
-import { Domain }              from '../src/Domain'
-import { HSL }                 from '../src/HSL'
-import { RGB }                 from '../src/RGB'
+import { indexedTo } from '@vect/object-mapper'
+import { Domain }    from '../resources/Domain.js'
+import { HSL }       from '../src/extends/HSL.js'
+import { RGB }    from '../src/extends/RGB.js'
 
 // XTerm's default colors