+++ packages/munsell/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @palett/munsell
 
+## 0.12.9
+
+### Patch Changes
+
+- fix bugs
+- Updated dependencies
+  - @palett/color-space@0.12.9
+  - @palett/convert@0.12.9
+
+## 0.12.8
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @palett/color-space@0.12.8
+  - @palett/convert@0.12.8
+
+## 0.12.7
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @palett/color-space@0.12.7
+  - @palett/convert@0.12.7
+
 ## 0.12.6
 
 ### Patch Changes